Add initial structure for projects, provide tests and documentation gen.
Includes:
carbon-resources Library project
carbon-resources CLI prject
Public API initial design - WIP with changes expected
Test project with tests for library and CLI
Documentation generation projects

NOTE:
This commit contains WIP work, designs are not final.
There are many todos in the code which highlight some of the main areas of ongoing design.
Yaml is being experimented with as a format for resourcelists, this may change.