Internalise resource prefix, resource prefix no longer exposed on public API
Resouces get a prefix such as res or app in the resourceGroups. This value is now
inferred from the resource type. New prefixes have been created for other file types
diff is a binary diff file
chunk is a bundle chunk
ResourceGroup is a resource group
BinaryGroup is a binary group
PatchGroup is a patch group
BundleGroup is a bundle group