Allow option of different resource sources for new files during patch application.
Update tests to use this new field.
Changed test data to use relative path style which are not expected to be compressed
This makes the test data correct and so the checksums match expected.
Tests now pass, however real patching is still not taking place yet.
Naive approach to patching and bundling basics functional.
Approach proves ResourceGroup filetype structure but will struggle with large files which is unnacceptable.