Fix crash when unzipping files larger than the chunk size
When calculating the amount of data to append, take into account the amount of data that has already been appended to the string.

The value used to decide the amount to append was the total length of the data that had been output so far. Subtracting the length of the output string should correct this to the amount of data that has not yet been appended to the string.