Add useful stats to ResourceGroups to track resource file sizes
This will be useful information for comparing patch size creation
As patches are are series of smaller patch files.
It is also useful to see the colllective size of a list.
This value can be compared to work out ratios of patch size/bundle size
compared to the input ResourceGroup