Make it possible to create and apply binary patches
This adds basic functionality around creating and applying binary patches. The generated patches do not yet have any compression applied, so we should expect them to be quite large. to see how good the patching is, we'll have to compress the resulting files, or add the compression/decompression as a part of the write/read functionality when creating the patches.

In order to get this working I hadd to add a couple of methods to the ChunkStream. These should help the reader when they don't want to read an entire chunk of data from the stream at once. I'm not sure if this is a good approach.