Fix issues preventing compilation on macOS
We need to bump our deployment target to 10.15 in order to be able to use std::filesystem::path. This should not be a problem, since that's the official minspec version.

Sinplify the definition of a couple of macros, making them into CMake defines rather than creating a file. The file was not being picked up by the macOS build.

Add definitions for exporting the API when not compiling under Windows.

Ensure consistent types for variables.