Fix bugs in streaming compression
Two problems here. Firstly, the amount available is not necessarily the same as the amount processed by zlib, so calculating that separatel. Secondly, We need to make sure to call deflate when we finish processing the buffer, or we will have unflushed data remaining in the buffer, resulting in Z_DATA_ERROR when calling deflateEnd.