Fix test failure on macOS due to OS specific assumptions
The binary operation parameter will be different based on platform, so we can't just use one file to determine if the output is ok. The order of exported YAML elements seems to be different between platforms. Depending on these coming out in the correct order is not ideal, and should be fixed at some point.