Make parameter versioning more specific
In the past, it was assumed that new parameters would only ever get added. This change allows us to keep track of which parameters get added in which version, as well as in which version they are deprecated. It also allows us to keep track of in which context a parameter is expected. This should help us out in the future if we end up supporting older versions.