Fix handling of default values when parsing arguments
When setting arguments as required and setting their default values, this was done on a copy of the argument, rather than on the argument used by the parser. Therefore, make sure to keep a reference to the argument in the parser and use that instead.

Fixing this would cause test failures in the CreatePatch cli operation test, since some of the parameters had been flagged as required. Since these same parameters had default values, and the test expected the command to work without them, these are no longer marked as required.

This removes the need for checking if a parameter was provided by the user before calling get, since now we get the default values when an explicit value has not been set.