Fix CLI verbosity setting and refactor CLI structure to match help
Since downgrading argparse it appears that specifying verbosity using
argument counting wansnt working, reading readme I can't find a reference
to it so it is likely not supported.

Changed it to be a standard optional argument that can be passed a number
does the same thing but isn't as slick.

Change also addresses and inconsistancy in usage compared to what the CLI
help was reporting as the 'usage'. Again this structure change was initially
due to the downgrade for argparse and no longer having subparsers.

Note: argparse was downgraded as we currently have agents that use v141 toolset
and later argparse versions caused internal compiler errors.