Generate indexes to speed up patch generation
Previously, the rolling checksum would be generated for the source file once per chunk, which would end up wasting a lot of time. Instead, generate the rolling checksum for each byte up front and save it in temporary files. This should significantly speed up the patch process for large files.

This saves us a lot of CPU time, but as a tradeoff the generated indexes take up 8x the size of the file being processed on disk. In order to minimize the file size, the amount of space taken up by the rolling checksum was reduced.