Speed up patch creation by reducing index size
Since we know in advance which parts of the destination file we are interested in, we can calculate the rolling checksums for those, and use them to filter the entries that we save out to the index files. This seems to result in a significant speedup as well as reducing the disk space required for the index files.