Optimize bundle creation
The bundle creation process was waiting until it had enough uncompressed data to produce a whole chunk before doing anything. This could cause a lot of memory being used when the files are large and compress very well. Therefore, refactor this to be more in line with the streaming paradigm, processing a bit at a time and outputting to files.

To prevent double work when processing the chunks, use the compressed and chunked files that were output from the streaming process.

By mistake, the patch process was using BundleStreamOut where it should have been using BundleStreamIn. At some point we should perform a similar optimization for the patch process, since that currently has the same problem when it comes to processing whole chunks at a time.

Deleting BundleStreamIn.h from the source folder, since the file does not seem to be used.

Cleaning up return value from << operator when creating a bundle. A return value of "false" now signifies failure, whereas the information about whether or not data has been returned has been moved to the GetChunk struct.

https://ccpgames.atlassian.net/browse/PLAT-9501