Include carbon-resources-static in CMake install rules
The launchers current node wrappers/integrations are set up for static libraries. Since we are already building one for the executable, simply include it in the install rules.

Configuring the carbon-resources-tools dependency as a BUILD_LOCAL_INTERFACE to keep it private. Without it CMake would raise an error complaining that carbon-resources-tools is not in any export set.